+++ src/App.tsx
@@ -1,0 +1,11 @@
+import Main from './Main';
+
+function App() {
+  return (
+    <div>
+      <Main />
+    </div>
+  );
+}
+
+export default App;
