--- src/index.js
@@ -1,16 +1,0 @@
-import { BrowserRouter} from 'react-router-dom';
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import "./index.css";
-
-
-
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-    <BrowserRouter>
-    <App />
-  </BrowserRouter>
-);
-
