--- src/pages/TestPage.js
@@ -1,11 +1,0 @@
-
-import PopOver from "../components/domains/folder/PopOver"
-
-function TestPage(){
-    return(
-        <div>
-           <PopOver/>
-        </div>
-    )
-}
-export default TestPage
