--- src/services/GetData.js
@@ -1,12 +1,0 @@
-import { API_BASE_URL } from "../constants";
-
-export async function getDate(url) {
-    try {
-      const response = await fetch(`${API_BASE_URL}${url}`);
-      if (!response.ok) throw new Error();
-      return await response.json();
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
