--- src/utils/AddFolder.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const AddFolder = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default AddFolder
