--- src/utils/AddToFolder.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const AddToFolder = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default AddToFolder
