--- src/utils/DeletFolder.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const deletFolder = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default deletFolder
