--- src/utils/DeletLink.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const DeletLink = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default DeletLink
