--- src/utils/SharedFolder.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const SharedFolder = () => {
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default SharedFolder
